user_id_is_biginteger

+++ aiogram_bot/database/models.py
@@ -52,7 +52,7 @@ class Chat(models.Model):
 
 
 class User(models.Model):
-    id = models.IntegerField(unique=True, primary_key=True)
+    id = models.BigIntegerField(unique=True, primary_key=True)
     username = models.CharField(max_length=32)
     is_subscription_complete = models.BooleanField(default=False)
